benchmark_framework/requirements.txt
PyYAML>=6.0
pyyaml
opencv-python
torch
torchvision
numpy
onnxruntime
psutil